+++ lib/interactive/responsive_text_field.dart
@@ -50,7 +50,7 @@ class ResponsiveTextField extends ResponsiveStatelessWidget {
     minLines,
     expands = false,
     maxLength,
-    maxLengthEnforcement = true,
+    maxLengthEnforcement,
     onChanged,
     onEditingComplete,
     onSubmitted,