+++ test/responsive_widgets_prefix_test.dart
@@ -1,7 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:responsive_widgets_prefix/responsive_widgets_prefix.dart';
-
 void main() {
   test('', () {